Make create and delete operations with BigInteger. Correct NPE bags in create and delete operations. Correct NPE bags in PersistenceFacade .

+++ src/main/java/com/netcracker/project/study/persistence/facade/Facade.java
@@ -5,6 +5,7 @@ import com.netcracker.project.study.model.Model;
 import com.netcracker.project.study.model.client.Client;
 import com.netcracker.project.study.persistence.PersistenceEntity;
 
+import java.math.BigInteger;
 import java.util.List;
 
 public interface Facade{
@@ -13,9 +14,9 @@ public interface Facade{
 
     void update(Model model) throws NoSuchFieldException, IllegalAccessException;
 
-    void delete(long objectId) throws NoSuchFieldException, IllegalAccessException;
+    void delete(BigInteger objectId) throws NoSuchFieldException, IllegalAccessException;
 
-    <T extends Model> T getOne(long objectId, Class modelClass) throws InstantiationException, IllegalAccessException;
+    <T extends Model> T getOne(BigInteger objectId, Class modelClass) throws InstantiationException, IllegalAccessException;
 
     List<? extends Model> getAll(long objectTypeId, Class modelClass) throws InstantiationException, IllegalAccessException;
 }
